+++ .github/CHANGELOG.md
@@ -9,6 +9,9 @@
 * Remove the old device API references in the Lightning repo and test suite.
   [(#1057)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1057)
 
+* Remove unnecessary adjoint pytest skip for Lightning-Kokkos.
+  [(#1048)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1048)
+
 * Update source code to use black formatter 25.1.0.
   [(#1059)](https://github.com/PennyLaneAI/pennylane-lightning/pull/1059)